Исправление багов

+++ src/main/java/SvistoPerdelki/Particles.java
@@ -1,6 +1,5 @@
 package SvistoPerdelki;
 
-import jdk.nashorn.internal.ir.Block;
 import org.bukkit.Location;
 import org.bukkit.Particle;
 import org.bukkit.World;
@@ -16,9 +15,9 @@ public class Particles {
 
         double step = 0.1D;
 
-        for (int x = 0; x < 3; x += step) {
-            for (int z = 0; z < 3; z += step) {
-                for (int y = 0; y <= 3; y += step) {
+        for (double x = 0; x < 3; x += step) {
+            for (double z = 0; z < 3; z += step) {
+                for (double y = 0; y <= 3; y += step) {
                     Location loc = new Location(world, startX + x, startY + y, startZ + z);
                     if (y != 3 && y != 0) {
                         if ((x >= 0 && z == 0) || (x >= 0 && z == 2) || (x == 0 && z >= 0) || (x == 2 && z >= 0))
